pyproject.toml
[project]
name = "claude-agents"
version = "0.1.0"
description = "AI agent mono-repo powered by Claude Agent SDK"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "playwright>=1.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.22.0",
    "httpx[http2]>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common", "agents"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
